+++ miniconf_derive/src/lib.rs
@@ -118,9 +118,8 @@ fn metadata_arm((i, struct_field): (usize, &StructField)) -> proc_macro2::TokenS
     if struct_field.deferred {
         quote! {
             #i => {
-                let mut meta = <#field_type>::metadata();
-                // Length of separator and field name
-                meta.max_length += 1 + stringify!(#field_name).len();
+                let mut meta = <#field_type>::metadata(separator_length);
+                meta.max_length += separator_length + stringify!(#field_name).len();
                 meta.max_depth += 1;
                 meta
             }
@@ -129,8 +128,7 @@ fn metadata_arm((i, struct_field): (usize, &StructField)) -> proc_macro2::TokenS
         quote! {
             #i => {
                 let mut meta = miniconf::Metadata::default();
-                // Length of separator and field name
-                meta.max_length = 1 + stringify!(#field_name).len();
+                meta.max_length = separator_length + stringify!(#field_name).len();
                 meta.max_depth = 1;
                 meta.count = 1;
                 meta
@@ -205,7 +203,7 @@ fn derive_struct(
                 }
             }
 
-            fn metadata() -> miniconf::Metadata {
+            fn metadata(separator_length: usize) -> miniconf::Metadata {
                 let mut meta = miniconf::Metadata::default();
 
                 for index in 0.. {