+++ tests/generics.rs
@@ -13,7 +13,7 @@ fn generic_type() {
     assert_eq!(settings.data, 3.0);
 
     // Test metadata
-    let metadata = Settings::<f32>::metadata();
+    let metadata = Settings::<f32>::metadata(1);
     assert_eq!(metadata.max_depth, 1);
     assert_eq!(metadata.max_length, "/data".len());
     assert_eq!(metadata.count, 1);
@@ -33,7 +33,7 @@ fn generic_array() {
     assert_eq!(settings.data[0], 3.0);
 
     // Test metadata
-    let metadata = Settings::<f32>::metadata();
+    let metadata = Settings::<f32>::metadata(1);
     assert_eq!(metadata.max_depth, 2);
     assert_eq!(metadata.max_length, "/data/0".len());
     assert_eq!(metadata.count, 2);
@@ -57,7 +57,7 @@ fn generic_struct() {
     assert_eq!(settings.inner.data, 3.0);
 
     // Test metadata
-    let metadata = Settings::<Inner>::metadata();
+    let metadata = Settings::<Inner>::metadata(1);
     assert_eq!(metadata.max_depth, 1);
     assert_eq!(metadata.max_length, "/inner".len());
     assert_eq!(metadata.count, 1);
@@ -83,7 +83,7 @@ fn generic_atomic() {
     assert_eq!(settings.atomic.inner[0], 3.0);
 
     // Test metadata
-    let metadata = Settings::<f32>::metadata();
+    let metadata = Settings::<f32>::metadata(1);
     assert_eq!(metadata.max_depth, 1);
     assert_eq!(metadata.max_length, "/atomic".len());
 }