+++ tests/structs.rs
@@ -34,7 +34,7 @@ fn atomic_struct() {
     assert_eq!(settings, expected);
 
     // Check that metadata is correct.
-    let metadata = Settings::metadata();
+    let metadata = Settings::metadata(1);
     assert_eq!(metadata.max_depth, 1);
     assert_eq!(metadata.max_length, "/c".len());
     assert_eq!(metadata.count, 3);
@@ -70,7 +70,7 @@ fn recursive_struct() {
     assert!(settings.set("/c", b"{\"a\": 5}").is_err());
 
     // Check that metadata is correct.
-    let metadata = Settings::metadata();
+    let metadata = Settings::metadata(1);
     assert_eq!(metadata.max_depth, 2);
     assert_eq!(metadata.max_length, "/c/a".len());
     assert_eq!(metadata.count, 3);